requirements.txt
langchain == 0.0.242
openai == 0.27.8
redis == 5.0.1
pandas == 2.0.3
sentence-transformers == 2.2.2
tiktoken == 0.5.1
streamlit == 1.27.2
python-dotenv = 1.0.0